+++ tests/Api/ServiceTest.php
@@ -1061,11 +1061,6 @@ class ServiceTest extends BaseTestCase
                 Entities\Layout::class,
             ],
             [
-                'leadSources',
-                Services\LeadSourceService::class,
-                Entities\LeadSource::class,
-            ],
-            [
                 'paymentCards',
                 Services\PaymentCardService::class,
                 Entities\PaymentCard::class,